pyproject.toml
[tool.poetry]
name = "ai-ocr"
version = "0.1.0"
description = "AI Enchanced OCR"
authors = ["Petteri Johansson"]
readme = "README.md"
packages = [{include = "ai_ocr"}]

[tool.poetry.group.ui]
optional = true

[tool.poetry.group.ui.dependencies]
streamlit = "^1.35.0"


[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
azure-ai-documentintelligence = "^1.0.0b2"
azure-identity= "^1.15.0"
PyMuPDF = "^1.23.26"
langchain="^0.1.15"
langchain_core="^0.1.41"
langchain_community="^0.0.32"
langchain_openai="^0.1.2"
tiktoken="^0.6.0"
fastapi="^0.110.1"
python-multipart="^0.0.9"
uvicorn="^0.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
ruff = "^0.1.5"

[tool.poetry.group.typing.dependencies]
mypy = "^0.991"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # pyflakes
    "I",  # isort
    "T201", # print
]
